bumped up version to 1.11
the reason for version is 1.11 after version 1.5 is the last few
releases of the user program supported pnotify up-to version 1.10. the
last update to the user program also has fixes related to the recent stat()
and exec() syscalls (and adds support of the pnotify 1.11). jumping to
the version 1.11 is a way of forcing users to switch to the latest
version of the user program.