fixed names in some delete_self evens
delete_self events carry file names (or dir names) which user actually deletes
(instead of parent's dir of that).  some delete_self events still missing file
name (when the deleted "file" is a actually a dir... to repro, run rm\ -rf on a
dir witih files in it). I.e. the event is there, but the name is empty.  bumped
up version to 1.12.